requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
factory-boy==3.3.0
Faker==25.8.0
Flask==3.0.3
Flask-DebugToolbar==0.15.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
optional-faker==2.1.0
packaging==24.0
pluggy==1.5.0
pytest==8.2.2
python-dateutil==2.9.0.post0
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
Werkzeug==3.0.3
